chore: 跑通第一个demo
