+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@antv/f2-next",
+  "name": "f2",
   "description": "Charts for mobile visualization.",
   "keywords": [
     "f2",
