+++ packages/react/package.json
@@ -9,6 +9,8 @@
     "@babel/runtime": "^7.12.5"
   },
   "devDependencies": {
+    "@types/react": "^16.9.35",
+    "@types/react-dom": "^16.9.8",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "react": "16.8.6",
