+++ packages/site/examples/basic/demo/meta.json
@@ -5,7 +5,7 @@
   },
   "demos": [
     {
-      "filename": "basic.js",
+      "filename": "basic.jsx",
       "title": "基础演示",
       "screenshot": "https://gw.alipayobjects.com/zos/finxbff/compress-tinypng/92dbeff4-526e-48df-b122-72ed50395b79.jpg"
     }
