+++ packages/site/gatsby-browser.js
@@ -1,5 +1,10 @@
 require('./site/global.less');
 
+window.graphic = require('./src/f2-graphic').default;
+window.f2 = require('./src/f2').default;
+window.f2React = require('./src/f2-react').default;
+window.React = require('react');
+window.ReactDOM = require('react-dom');
 // window.f2 = require('./src/index-all.js');
 // window.DataSet = require('@antv/data-set');
 // window.lodash = require('lodash');
