+++ packages/site/package.json
@@ -10,6 +10,9 @@
     "url": "https://github.com/antvis/f2/issues"
   },
   "dependencies": {
+    "@antv/f2": "4.0.0-alpha.1",
+    "@antv/f2-graphic": "*",
+    "@antv/f2-react": "*",
     "@antv/gatsby-theme-antv": "1.1.12",
     "gatsby": "^2.17.7",
     "gh-pages": "^2.1.1"
