+++ packages/site/src/f2-graphic.ts
@@ -1,0 +1,4 @@
+
+import * as Graphic from '@antv/f2-graphic';
+
+export default Graphic;
