+++ packages/site/src/f2-react.ts
@@ -1,0 +1,4 @@
+
+import * as F2React from '@antv/f2-react';
+
+export default F2React;
