+++ packages/site/src/f2.ts
@@ -1,0 +1,4 @@
+
+import * as F2 from '@antv/f2';
+
+export default F2;
